+++ api/V0GetKubeconfig.go
@@ -16,7 +16,7 @@ func V0GetKubeconfig(w http.ResponseWriter, r *http.Request) {
 	logrus.Debug("HTTP GET V0GetKubeconfig")
 
 	client := &http.Client{}
-	req, _ := http.NewRequest("GET", "http://"+config.K3SServerAPIHostname+":"+strconv.Itoa(config.K3SServerAPIPort)+"/api/k3s/v0/config", nil)
+	req, _ := http.NewRequest("GET", "http://"+config.M3SBootstrapServerHostname+":"+strconv.Itoa(config.M3SBootstrapServerPort)+"/api/k3s/v0/config", nil)
 	res, err := client.Do(req)
 
 	if err != nil {
@@ -37,7 +37,7 @@ func V0GetKubeconfig(w http.ResponseWriter, r *http.Request) {
 	}
 
 	// replace the localhost server string with the mesos agent hostname and dynamic port
-	destUrl := config.K3SServerAPIHostname + ":" + strconv.Itoa(config.K3SServerPort)
+	destUrl := config.M3SBootstrapServerHostname + ":" + strconv.Itoa(config.K3SServerPort)
 	kubconf := strings.Replace(string(content), "127.0.0.1:6443", destUrl, -1)
 
 	w.WriteHeader(http.StatusAccepted)