+++ mesos_cli/m3s/main.py
@@ -19,11 +19,13 @@ The m3s plugin.
 """
 
 from cli.exceptions import CLIException
-from cli.mesos import get_tasks
 from cli.plugins import PluginBase
 from cli.util import Table
 
 from cli.mesos import TaskIO
+from cli import http
+from cli.exceptions import CLIException
+
 
 PLUGIN_NAME = "m3s"
 PLUGIN_CLASS = "M3s"
@@ -32,6 +34,33 @@ VERSION = "0.1.0"
 
 SHORT_HELP = "Interacts with the Kubernetes Framework M3s"
 
+def get_framework_address(framework_id, master, config):
+    """
+    Given a master and an framework id, return the framework address
+    by checking the /master/frameworks endpoint of the master.
+    """
+    frameworks = get_frameworks(master, config)
+
+    for framework in frameworks:
+        if framework["id"] == framework_id:
+            return framework["webui_url"]
+    raise CLIException("Unable to find framework '{id}'".format(id=framework_id))
+
+
+def get_frameworks(master, config):
+    """
+    Get the frameworks in a Mesos cluster.
+    """
+    endpoint = "master/frameworks/"
+    keyword="frameworks"
+    try:
+        data = http.get_json(master, endpoint, config)["frameworks"]
+    except Exception as exception:
+        raise CLIException(
+            "Could not open '/{endpoint}' on master: {error}"
+            .format(endpoint=endpoint, error=exception))
+
+    return data
 
 class M3s(PluginBase):
     """
@@ -40,16 +69,72 @@ class M3s(PluginBase):
 
     COMMANDS = {
         "kubeconfig": {
-            "arguments": ['<framwork-id>'],
+            "arguments": ['<framework-id>'],
             "flags": {},
             "short_help": "Get kubernetes configuration file",
             "long_help": "Get kubernetes configuration file"
+        },
+        "list": {
+            "arguments": [],
+            "flags": {
+                "-a --all": "list all M3s frameworks, not only running [default: False]"
+            },
+            "short_help": "Show list of running M3s frameworks",
+            "long_help": "Show list of running M3s frameworks",
         }
     }
-
 
     def kubeconfig(self, argv):
         """
         Load the kubeconfig file from the kubernetes api server
         """
+
+        try:
+            master = self.config.master()
+            config = self.config
+        except Exception as exception:
+            raise CLIException("Unable to get leading master address: {error}"
+                               .format(error=exception))
+
+        framework_address = get_framework_address(argv["<framework-id>"], master, config)
+        data = http.read_endpoint(framework_address, "/v0/server/config", config)
+
+        print (data)
+
+
+    def list(self, argv):
+        """
+        Show a list of running M3s frameworks
+        """
+
+        try:
+            master = self.config.master()
+            config = self.config
+        except Exception as exception:
+            raise CLIException("Unable to get leading master address: {error}"
+                               .format(error=exception))
+
+        data = get_frameworks(master, config)
+        try:
+            table = Table(["ID", "Active", "WebUI", "Name"])
+            for framework in data:
+                if (not argv["--all"] and framework["active"] is not True) or "m3s" not in framework["name"].lower():
+                    continue
+
+                active = "False"
+                if framework["active"]:
+                    active = "True"
+
+                table.add_row([framework["id"],
+                               active,
+                               framework["webui_url"],
+                               framework["name"]])
+        except Exception as exception:
+            raise CLIException("Unable to build table of M3s frameworks: {error}"
+                               .format(error=exception))
+
+        print(str(table))
+        
+
+